Fixes api.js accessing an uninitialized flowCardManager in app.js
Moves up the initialization of flowCardManager in app.js onInit before
any async calls are made. Also added checks in the API endpoints that
check if flowCardManager is initialized, and if not will wait for 500ms.
A bit hacky, but flowCardManager is a sync constructor that cannot be
awaited.

Related Sentry issue: https://sentry.io/athom-bv/comifttt/issues/823580276/?query=is%3Aunresolved